send image

+++ src/main/java/com/mproduits/model/Product.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)  // Use IDENTITY for H2 auto-increment
-    private int id;
+    private Long id;
     private String titre;
     private String description;
     private String image;  // Path to the image file
@@ -19,20 +19,22 @@ public class Product {
     public Product() {}
 
     // Parameterized constructor
-    public Product(int id, String titre, String description, String image, Double prix) {
+    public Product(Long id, String titre, String description, String image, Double prix) {
         this.id = id;
         this.titre = titre;
         this.description = description;
         this.image = image;
         this.prix = prix;
+
     }
 
+
     // Getters and Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
